Move GetCheckFromActor to randomizer_identify_actor
